+++ src/main/java/io/electrum/moneytransfer/model/MoneyTransferRedeemRequest.java
@@ -5,15 +5,17 @@ import java.util.Objects;
 import javax.validation.Valid;
 import javax.validation.constraints.NotNull;
 
+import org.hibernate.validator.constraints.Length;
+
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 import io.electrum.vas.Utils;
+import io.electrum.vas.model.Amounts;
 import io.electrum.vas.model.EncryptedPin;
 import io.electrum.vas.model.LedgerAmount;
 import io.electrum.vas.model.Transaction;
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
-import org.hibernate.validator.constraints.Length;
 
 /**
  * Used to submit data in a call to the redeemOrder operation.
@@ -33,18 +35,22 @@ public class MoneyTransferRedeemRequest extends Transaction {
 
    private String cashierComment = null;
 
+   private Amounts amounts = null;
+
    public MoneyTransferRedeemRequest amount(LedgerAmount amount) {
       this.amount = amount;
       return this;
    }
 
    /**
-    * Get amount
+    * The amount to be transferred. This field may be deprecated in a future version of the API. We encourage you to
+    * please also populate the 'amounts.requestAmount' field with this information.
     *
     * @return amount
     **/
    @JsonProperty("amount")
-   @ApiModelProperty(value = "")
+   @ApiModelProperty(value = "The amount to be transferred. This field may be deprecated in a future version of the API. " +
+         "We encourage you to please also populate the 'amounts.requestAmount' field with this information.")
    @Valid
    public LedgerAmount getAmount() {
       return amount;
@@ -161,6 +167,31 @@ public class MoneyTransferRedeemRequest extends Transaction {
       this.cashierComment = cashierComment;
    }
 
+   public MoneyTransferRedeemRequest amounts(Amounts amounts) {
+      this.amounts = amounts;
+      return this;
+   }
+
+   /**
+    * Amounts which make up the transaction. The existing 'amount' field currently takes precedence over this 'amounts'
+    * field, however the use of this 'amounts' field is encouraged. The 'amount' field may be deprecated in a future
+    * version of this API.
+    *
+    * @return amounts
+    **/
+   @JsonProperty("amounts")
+   @ApiModelProperty(value = "Amounts which make up the transaction. The existing 'amount' field currently takes "
+         + "precedence over this 'amounts' field, however the use of this 'amounts' field is encouraged. The 'amount' "
+         + "field may be deprecated in a future version of this API.")
+   @Valid
+   public Amounts getAmounts() {
+      return amounts;
+   }
+
+   public void setAmounts(Amounts amounts) {
+      this.amounts = amounts;
+   }
+
    @Override
    public boolean equals(Object o) {
       if (this == o)
@@ -173,12 +204,12 @@ public class MoneyTransferRedeemRequest extends Transaction {
       return Objects.equals(amount, that.amount) && Objects.equals(pin, that.pin)
             && Objects.equals(recipientDetails, that.recipientDetails)
             && Objects.equals(orderRedeemRef, that.orderRedeemRef)
-            && Objects.equals(cashierComment, that.cashierComment);
+            && Objects.equals(cashierComment, that.cashierComment) && Objects.equals(amounts, that.amounts);
    }
 
    @Override
    public int hashCode() {
-      return Objects.hash(super.hashCode(), amount, pin, recipientDetails, orderRedeemRef, cashierComment);
+      return Objects.hash(super.hashCode(), amount, pin, recipientDetails, orderRedeemRef, cashierComment, amounts);
    }
 
    @Override
@@ -200,6 +231,7 @@ public class MoneyTransferRedeemRequest extends Transaction {
       sb.append("    orderRedeemRef: ").append(Utils.toIndentedString(orderRedeemRef)).append("\n");
       sb.append("    customerProfileId: ").append(Utils.toIndentedString(customerProfileId)).append("\n");
       sb.append("    cashierComment: ").append(Utils.toIndentedString(cashierComment)).append("\n");
+      sb.append("    amounts: ").append(Utils.toIndentedString(amounts)).append("\n");
       sb.append("}");
       return sb.toString();
    }