+++ src/docs/devguide/hugo/site/content/release-notes.md
@@ -6,6 +6,8 @@ Released xx Dec 2019
 
 - Updated the `MoneyTransferAuthResponse` and `MoneyTransferAuthRequest` models with a new `amounts` field:
     - The `amounts` field allows for more amounts to be used to describe a Money Transfer transaction.
+        - The `amount` field takes precedence over this `amounts` field, however the use of this 
+        `amounts` field is encouraged. The `amount` field may be deprecated in a future version of this API.
 
 ## v2.17.0